+++ src/routes/event.js
@@ -1,4 +1,4 @@
-const findEvent = require('../actions/find_event.js');
+const findEvent = require('../actions/get_event.js');
 
 module.exports = (req, res) => {
   findEvent(req.params.id, 'event', res);
