+++ src/routes/news.js
@@ -1,4 +1,4 @@
-const findNews = require('../actions/find_news.js');
+const findNews = require('../actions/get_news.js');
 
 module.exports = (req, res) => {
   findNews(req.params.id, 'news', res);
